requirements.txt
tensorflow
pandas 
dvc
notebook
numpy
matplotlib
seaborn
python-box==6.0.2
pyYAML
tqdm
ensure==1.0.2
joblib
types-PyYAML
scipy
Flask
Flask-Cors
mlflow==2.5.0
python-dotenv==1.0.0
-e . # Works to install the setup.py file